cncb主页数据格式矫正

+++ src/main/java/cn/ac/big/bigd/webservice/mapper/gsa/GsaMapper.java
@@ -31,7 +31,7 @@ public interface GsaMapper {
     int getJournalCounts();
     CraDownLoad selectCraByAccession(String accession);
     CraDownLoad selectCraByProAccession(String accession);
-    int selectMaxPb();
+    Double selectMaxPb();
     List<DateFileSize> selectAllDate();
 
     int getCraCountByPrjIdNoDelete(int prjId);
